[Fix] 카테고리 페이지에서 이미지를 불러오지 못하는 버그 해결 #166
